+++ src/main/java/org/betonquest/betonquest/item/typehandler/BookHandler.java
@@ -2,8 +2,6 @@ package org.betonquest.betonquest.item.typehandler;
 
 import org.betonquest.betonquest.config.Config;
 import org.betonquest.betonquest.exceptions.InstructionParseException;
-import org.betonquest.betonquest.item.QuestItem;
-import org.betonquest.betonquest.item.QuestItem.Existence;
 import org.betonquest.betonquest.utils.Utils;
 import org.bukkit.ChatColor;
 import org.bukkit.inventory.meta.BookMeta;
@@ -13,7 +11,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
-@SuppressWarnings({"PMD.CommentRequired", "PMD.GodClass"})
+@SuppressWarnings("PMD.CommentRequired")
 public class BookHandler implements ItemMetaHandler<BookMeta> {
     private String title = Config.getMessage(Config.getLanguage(), "unknown_title");
 
@@ -92,22 +90,20 @@ public class BookHandler implements ItemMetaHandler<BookMeta> {
 
     @Override
     public void populate(final BookMeta bookMeta) {
-        bookMeta.setTitle(getTitle());
-        bookMeta.setAuthor(getAuthor());
-        bookMeta.setPages(getText());
+        bookMeta.setTitle(title);
+        bookMeta.setAuthor(author);
+        bookMeta.setPages(text);
     }
 
     @Override
     public boolean check(final BookMeta bookMeta) {
-        return checkTitle(bookMeta.getTitle()) && checkAuthor(bookMeta.getAuthor()) && checkText(bookMeta.getPages());
-    }
-
-    public String getTitle() {
-        return title;
+        return checkExistence(titleE, title, bookMeta.getTitle())
+                && checkExistence(authorE, author, bookMeta.getAuthor())
+                && checkText(bookMeta.getPages());
     }
 
     public void setTitle(final String string) {
-        if (QuestItem.NONE_KEY.equalsIgnoreCase(string)) {
+        if (Existence.NONE_KEY.equalsIgnoreCase(string)) {
             titleE = Existence.FORBIDDEN;
         } else {
             title = string.replace('_', ' ');
@@ -116,12 +112,8 @@ public class BookHandler implements ItemMetaHandler<BookMeta> {
         }
     }
 
-    public String getAuthor() {
-        return author;
-    }
-
     public void setAuthor(final String string) {
-        if (QuestItem.NONE_KEY.equalsIgnoreCase(string)) {
+        if (Existence.NONE_KEY.equalsIgnoreCase(string)) {
             authorE = Existence.FORBIDDEN;
         } else {
             author = string.replace("_", " ");
@@ -129,12 +121,8 @@ public class BookHandler implements ItemMetaHandler<BookMeta> {
         }
     }
 
-    public List<String> getText() {
-        return text;
-    }
-
     public void setText(final String string) {
-        if (QuestItem.NONE_KEY.equalsIgnoreCase(string)) {
+        if (Existence.NONE_KEY.equalsIgnoreCase(string)) {
             text.add(""); // this will prevent "Invalid book tag" message in the empty book
             textE = Existence.FORBIDDEN;
         } else {
@@ -142,14 +130,6 @@ public class BookHandler implements ItemMetaHandler<BookMeta> {
             text.replaceAll(textToTranslate -> ChatColor.translateAlternateColorCodes('&', textToTranslate));
             textE = Existence.REQUIRED;
         }
-    }
-
-    public boolean checkTitle(@Nullable final String string) {
-        return checkExistence(titleE, title, string);
-    }
-
-    public boolean checkAuthor(@Nullable final String string) {
-        return checkExistence(authorE, author, string);
     }
 
     private boolean checkExistence(final Existence existence, @Nullable final String present, @Nullable final String string) {