+++ src/main/java/org/betonquest/betonquest/item/typehandler/CustomModelDataHandler.java
@@ -1,7 +1,6 @@
 package org.betonquest.betonquest.item.typehandler;
 
 import org.betonquest.betonquest.exceptions.InstructionParseException;
-import org.betonquest.betonquest.item.QuestItem;
 import org.bukkit.inventory.meta.ItemMeta;
 import org.jetbrains.annotations.Nullable;
 
@@ -9,7 +8,7 @@ import java.util.Set;
 
 @SuppressWarnings("PMD.CommentRequired")
 public class CustomModelDataHandler implements ItemMetaHandler<ItemMeta> {
-    private QuestItem.Existence existence = QuestItem.Existence.WHATEVER;
+    private Existence existence = Existence.WHATEVER;
 
     private int modelData;
 
@@ -35,60 +34,38 @@ public class CustomModelDataHandler implements ItemMetaHandler<ItemMeta> {
         return null;
     }
 
-    public void parse(final String data) throws InstructionParseException {
-        try {
-            require(Integer.parseInt(data));
-        } catch (final NumberFormatException e) {
-            throw new InstructionParseException("Could not parse custom model data value: " + data, e);
-        }
-    }
-
-    public void require(final int customModelDataId) {
-        this.existence = QuestItem.Existence.REQUIRED;
-        this.modelData = customModelDataId;
-    }
-
-    public void forbid() {
-        this.existence = QuestItem.Existence.FORBIDDEN;
-    }
-
-    public QuestItem.Existence getExistence() {
-        return existence;
-    }
-
-    public boolean has() {
-        return existence == QuestItem.Existence.REQUIRED;
-    }
-
-    public int get() {
-        return modelData;
-    }
-
     @Override
     public void set(final String key, final String data) throws InstructionParseException {
         switch (key) {
-            case "custom-model-data" -> parse(data);
-            case "no-custom-model-data" -> forbid();
+            case "custom-model-data" -> {
+                try {
+                    this.existence = Existence.REQUIRED;
+                    this.modelData = Integer.parseInt(data);
+                } catch (final NumberFormatException e) {
+                    throw new InstructionParseException("Could not parse custom model data value: " + data, e);
+                }
+            }
+            case "no-custom-model-data" -> this.existence = Existence.FORBIDDEN;
             default -> throw new InstructionParseException("Unknown custom model data key: " + key);
         }
     }
 
     @Override
     public void populate(final ItemMeta meta) {
-        if (getExistence() == QuestItem.Existence.REQUIRED) {
-            meta.setCustomModelData(get());
+        if (existence == Existence.REQUIRED) {
+            meta.setCustomModelData(modelData);
         }
     }
 
     @Override
     public boolean check(final ItemMeta data) {
-        return existence == QuestItem.Existence.WHATEVER
-                || existence == QuestItem.Existence.FORBIDDEN && !data.hasCustomModelData()
-                || existence == QuestItem.Existence.REQUIRED && data.hasCustomModelData() && modelData == data.getCustomModelData();
+        return existence == Existence.WHATEVER
+                || existence == Existence.FORBIDDEN && !data.hasCustomModelData()
+                || existence == Existence.REQUIRED && data.hasCustomModelData() && modelData == data.getCustomModelData();
     }
 
     @Override
     public String toString() {
-        return existence == QuestItem.Existence.REQUIRED ? "custom-model-data:" + modelData : "";
+        return existence == Existence.REQUIRED ? "custom-model-data:" + modelData : "";
     }
 }