+++ src/main/java/org/betonquest/betonquest/item/typehandler/EnchantmentsHandler.java
@@ -1,9 +1,6 @@
 package org.betonquest.betonquest.item.typehandler;
 
 import org.betonquest.betonquest.exceptions.InstructionParseException;
-import org.betonquest.betonquest.item.QuestItem;
-import org.betonquest.betonquest.item.QuestItem.Existence;
-import org.betonquest.betonquest.item.QuestItem.Number;
 import org.betonquest.betonquest.utils.Utils;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.meta.EnchantmentStorageMeta;
@@ -23,6 +20,9 @@ public class EnchantmentsHandler implements ItemMetaHandler<ItemMeta> {
 
     private Existence checkersE = Existence.WHATEVER;
 
+    /**
+     * If the Enchantment need to be exact the same or just contain all specified enchantments.
+     */
     private boolean exact = true;
 
     public EnchantmentsHandler() {
@@ -65,7 +65,7 @@ public class EnchantmentsHandler implements ItemMetaHandler<ItemMeta> {
     public void set(final String key, final String data) throws InstructionParseException {
         switch (key) {
             case "enchants" -> set(data);
-            case "enchants-containing" -> setNotExact();
+            case "enchants-containing" -> exact = false;
             default -> throw new InstructionParseException("Unknown enchantment key: " + key);
         }
     }
@@ -97,7 +97,7 @@ public class EnchantmentsHandler implements ItemMetaHandler<ItemMeta> {
 
     public void set(final String enchants) throws InstructionParseException {
         final String[] parts = HandlerUtil.getNNSplit(enchants, "Enchantment is null!", ",");
-        if (QuestItem.NONE_KEY.equalsIgnoreCase(parts[0])) {
+        if (Existence.NONE_KEY.equalsIgnoreCase(parts[0])) {
             checkersE = Existence.FORBIDDEN;
             return;
         }
@@ -107,10 +107,6 @@ public class EnchantmentsHandler implements ItemMetaHandler<ItemMeta> {
             checkers.add(checker);
         }
         checkersE = Existence.REQUIRED;
-    }
-
-    public void setNotExact() {
-        exact = false;
     }
 
     public Map<Enchantment, Integer> get() {