+++ src/main/java/org/betonquest/betonquest/item/typehandler/FireworkEffectHandler.java
@@ -1,8 +1,6 @@
 package org.betonquest.betonquest.item.typehandler;
 
 import org.betonquest.betonquest.exceptions.InstructionParseException;
-import org.betonquest.betonquest.item.QuestItem;
-import org.betonquest.betonquest.item.QuestItem.Existence;
 import org.betonquest.betonquest.utils.Utils;
 import org.bukkit.Color;
 import org.bukkit.DyeColor;
@@ -57,7 +55,7 @@ public class FireworkEffectHandler {
         if (parts.length != 5) {
             throw new InstructionParseException("Incorrect effect format: " + string);
         }
-        if (QuestItem.NONE_KEY.equalsIgnoreCase(parts[1])) {
+        if (Existence.NONE_KEY.equalsIgnoreCase(parts[1])) {
             mainE = Existence.FORBIDDEN;
         } else if ("?".equals(parts[1])) {
             mainE = Existence.WHATEVER;
@@ -69,7 +67,7 @@ public class FireworkEffectHandler {
                 mainColors.add(fireworkColor == null ? regularColor : fireworkColor.getFireworkColor());
             }
         }
-        if (QuestItem.NONE_KEY.equalsIgnoreCase(parts[2])) {
+        if (Existence.NONE_KEY.equalsIgnoreCase(parts[2])) {
             fadeE = Existence.FORBIDDEN;
         } else if ("?".equals(parts[2])) {
             fadeE = Existence.WHATEVER;
@@ -189,5 +187,4 @@ public class FireworkEffectHandler {
                 return false;
         }
     }
-
 }