+++ src/main/java/org/betonquest/betonquest/item/typehandler/FlagHandler.java
@@ -1,7 +1,6 @@
 package org.betonquest.betonquest.item.typehandler;
 
 import org.betonquest.betonquest.exceptions.InstructionParseException;
-import org.betonquest.betonquest.item.QuestItem;
 import org.bukkit.inventory.ItemFlag;
 import org.bukkit.inventory.meta.ItemMeta;
 import org.jetbrains.annotations.Nullable;
@@ -22,7 +21,7 @@ public class FlagHandler implements ItemMetaHandler<ItemMeta> {
     /**
      * Existence of the flags.
      */
-    private QuestItem.Existence existence = QuestItem.Existence.WHATEVER;
+    private Existence existence = Existence.WHATEVER;
 
     /**
      * Construct a new FlagHandler.
@@ -50,58 +49,30 @@ public class FlagHandler implements ItemMetaHandler<ItemMeta> {
         return "flags:" + String.join(",", meta.getItemFlags().stream().map(ItemFlag::name).sorted().toList());
     }
 
-    /**
-     * Parse a String into the Set of ItemFlags.
-     *
-     * @param data The Set of ItemFlags.
-     * @throws InstructionParseException If there is an error parsing.
-     */
-    public void parse(final String data) throws InstructionParseException {
-        set(Arrays.stream(data.split(",")).map(ItemFlag::valueOf).collect(Collectors.toSet()));
-    }
-
-    /**
-     * Set the Set of ItemFlags in this handler.
-     *
-     * @param itemFlags The ItemFlags, or null if not set.
-     * @throws InstructionParseException If there is an error setting the flags.
-     */
-    public void set(@Nullable final Set<ItemFlag> itemFlags) throws InstructionParseException {
-        if (itemFlags == null || itemFlags.isEmpty()) {
-            this.itemFlags = Set.of();
-            this.existence = QuestItem.Existence.FORBIDDEN;
-        } else {
-            this.itemFlags = Set.copyOf(itemFlags);
-            this.existence = QuestItem.Existence.REQUIRED;
-        }
-    }
-
-    /**
-     * Get the Set of ItemFlags.
-     *
-     * @return The Set of ItemFlags.
-     */
-    public Set<ItemFlag> get() {
-        return itemFlags;
-    }
-
     @Override
     public void set(final String key, final String data) throws InstructionParseException {
         if (!"flags".equals(key)) {
             throw new InstructionParseException("Invalid flag key: " + key);
         }
-        parse(data);
+        final Set<ItemFlag> flags = Arrays.stream(data.split(",")).map(ItemFlag::valueOf).collect(Collectors.toSet());
+        if (flags.isEmpty()) {
+            this.itemFlags = Set.of();
+            this.existence = Existence.FORBIDDEN;
+        } else {
+            this.itemFlags = Set.copyOf(flags);
+            this.existence = Existence.REQUIRED;
+        }
     }
 
     @Override
     public void populate(final ItemMeta meta) {
-        get().forEach(meta::addItemFlags);
+        itemFlags.forEach(meta::addItemFlags);
     }
 
     @Override
     public boolean check(final ItemMeta data) {
-        return existence == QuestItem.Existence.WHATEVER
-                || existence == QuestItem.Existence.FORBIDDEN && data.getItemFlags().isEmpty()
-                || existence == QuestItem.Existence.REQUIRED && !data.getItemFlags().isEmpty() && itemFlags.equals(data.getItemFlags());
+        return existence == Existence.WHATEVER
+                || existence == Existence.FORBIDDEN && data.getItemFlags().isEmpty()
+                || existence == Existence.REQUIRED && !data.getItemFlags().isEmpty() && itemFlags.equals(data.getItemFlags());
     }
 }