+++ src/main/java/org/betonquest/betonquest/item/typehandler/HeadHandler.java
@@ -4,7 +4,6 @@ import io.papermc.lib.PaperLib;
 import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.api.profiles.Profile;
 import org.betonquest.betonquest.exceptions.InstructionParseException;
-import org.betonquest.betonquest.item.QuestItem;
 import org.betonquest.betonquest.utils.PlayerConverter;
 import org.bukkit.Bukkit;
 import org.bukkit.OfflinePlayer;
@@ -20,7 +19,6 @@ import java.util.stream.Collectors;
 /**
  * Handles metadata about player Skulls.
  */
-@SuppressWarnings("PMD.TooManyMethods")
 public abstract class HeadHandler implements ItemMetaHandler<SkullMeta> {
     /**
      * Owner metadata about the Skull.
@@ -45,12 +43,12 @@ public abstract class HeadHandler implements ItemMetaHandler<SkullMeta> {
     /**
      * Existence of the player UUID.
      */
-    private QuestItem.Existence playerIdE = QuestItem.Existence.WHATEVER;
+    private Existence playerIdE = Existence.WHATEVER;
 
     /**
      * Existence of the encoded texture.
      */
-    private QuestItem.Existence textureE = QuestItem.Existence.WHATEVER;
+    private Existence textureE = Existence.WHATEVER;
 
     /**
      * An optional player name owner of the skull.
@@ -73,7 +71,7 @@ public abstract class HeadHandler implements ItemMetaHandler<SkullMeta> {
     /**
      * Existence of the owner.
      */
-    private QuestItem.Existence ownerE = QuestItem.Existence.WHATEVER;
+    private Existence ownerE = Existence.WHATEVER;
 
     /**
      * Construct a new HeadHandler.
@@ -135,9 +133,22 @@ public abstract class HeadHandler implements ItemMetaHandler<SkullMeta> {
     @Override
     public void set(final String key, final String data) throws InstructionParseException {
         switch (key) {
-            case META_OWNER -> setOwner(data);
-            case META_PLAYER_ID -> setPlayerId(data);
-            case META_TEXTURE -> setTexture(data);
+            case META_OWNER -> {
+                if (Existence.NONE_KEY.equalsIgnoreCase(data)) {
+                    ownerE = Existence.FORBIDDEN;
+                } else {
+                    owner = data;
+                    ownerE = Existence.REQUIRED;
+                }
+            }
+            case META_PLAYER_ID -> {
+                this.playerId = UUID.fromString(data);
+                this.playerIdE = Existence.REQUIRED;
+            }
+            case META_TEXTURE -> {
+                this.texture = data;
+                this.textureE = Existence.REQUIRED;
+            }
             default -> throw new InstructionParseException("Unknown head key: " + key);
         }
     }
@@ -146,20 +157,6 @@ public abstract class HeadHandler implements ItemMetaHandler<SkullMeta> {
     @Override
     public void populate(final SkullMeta meta) {
         throw new UnsupportedOperationException("Use #populate(SkullMeta, Profile) instead");
-    }
-
-    /**
-     * Set the owner name to the specified value.
-     *
-     * @param string The new String name for the owner.
-     */
-    public void setOwner(final String string) {
-        if (QuestItem.NONE_KEY.equalsIgnoreCase(string)) {
-            ownerE = QuestItem.Existence.FORBIDDEN;
-        } else {
-            owner = string;
-            ownerE = QuestItem.Existence.REQUIRED;
-        }
     }
 
     /**
@@ -192,16 +189,6 @@ public abstract class HeadHandler implements ItemMetaHandler<SkullMeta> {
     }
 
     /**
-     * Set the player UUID to the specified value.
-     *
-     * @param playerId The new UUID player ID.
-     */
-    public void setPlayerId(final String playerId) {
-        this.playerId = UUID.fromString(playerId);
-        this.playerIdE = QuestItem.Existence.REQUIRED;
-    }
-
-    /**
      * Get the encoded texture.
      *
      * @return The encoded texture.
@@ -209,16 +196,6 @@ public abstract class HeadHandler implements ItemMetaHandler<SkullMeta> {
     @Nullable
     public String getTexture() {
         return texture;
-    }
-
-    /**
-     * Set the encoded texture to the specified value.
-     *
-     * @param texture The new encoded texture.
-     */
-    public void setTexture(final String texture) {
-        this.texture = texture;
-        this.textureE = QuestItem.Existence.REQUIRED;
     }
 
     /**