+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Removed
 ### Fixed
 - Reloading plugin did not reload Menu config
+- potions generated not extended/upgraded since 2.2.1
 ### Security
 
 ## [2.2.1] - 2025-01-12